Make overview retrieve data from repository
